55: extract `<Home>` and its test into separate files
