+++ frontend/src/App.test.tsx
@@ -15,7 +15,7 @@ import { createMemoryHistory } from "history";
 import { Router } from "react-router-dom";
 
 import { INITIAL_STATE, rootReducer } from "./store";
-import { NavigationBar, Home, About, Account } from "./App";
+import { NavigationBar, About, Account } from "./App";
 import { Alerts } from "./features/alerts/Alerts";
 
 // 2
@@ -42,35 +42,6 @@ import {
 
 // const BIG_VALUE_FOR_TIMEOUT_OF_ASYNCHRONOUS_OPERATIONS: number = 5 * 60 * 1000;
 // jest.setTimeout(BIG_VALUE_FOR_TIMEOUT_OF_ASYNCHRONOUS_OPERATIONS);
-
-describe("<Home>", () => {
-  test("renders a 'Welcome to VocabTreasury!' message", () => {
-    render(<Home />);
-    const headingElement = screen.getByText("Welcome to VocabTreasury!");
-
-    /*
-    The following statement throws a
-    `TypeError: expect(...).toBeInTheDocument is not a function`
-
-    The post and comments on
-    https://stackoverflow.com/questions/56547215/react-testing-library-why-is-tobeinthedocument-not-a-function
-    explain that:
-
-    - the reason for the error is that
-      `toBeInTheDocument` is not part of the React Testing Library
-
-    - the problem can be rectified
-      by adding `import '@testing-library/jest-dom' into this file
-
-    - in fact, the problem had been solved to begin with
-      by the Create React App utility itself,
-      because it had added the above-mentioned import statement into `setupTests.ts`
-      (which means that the "21: remove files and boilerplate code, which ..." commit was
-      what gave rise to this problem in this repository)
-    */
-    expect(headingElement).toBeInTheDocument();
-  });
-});
 
 describe("<NavigationBar>", () => {
   const alwaysVisibleLinks = ["VocabTreasury", "Home", "About"];
