+++ frontend/src/features/Home.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+export const Home = () => {
+  console.log(`${new Date().toISOString()} - React is rendering <Home>`);
+
+  return (
+    <React.Fragment>
+      {"<Home>"}
+      <div>
+        <h1>Welcome to VocabTreasury!</h1>
+      </div>
+    </React.Fragment>
+  );
+};
